pyproject.toml
[project]
name = "kakigoori"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgiref==3.8.1",
    "black==24.4.2",
    "boto3==1.35.15",
    "botocore==1.35.15",
    "certifi==2024.8.30",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "django==5.2",
    "django-stubs>=5.2.0",
    "gunicorn==23.0.0",
    "idna==3.8",
    "jmespath==1.0.1",
    "mypy-extensions==1.0.0",
    "packaging==23.2",
    "pathspec==0.12.1",
    "pillow==10.4.0",
    "platformdirs==4.2.2",
    "python-dateutil==2.9.0.post0",
    "requests==2.32.3",
    "s3transfer==0.10.2",
    "sentry-sdk==2.14.0",
    "six==1.16.0",
    "sqlparse==0.5.1",
    "tzdata>=2025.2",
    "urllib3==2.2.2",
]
